Properly implement jumpToItem method (#14)
Prior to this commit the jumpToItem method would
only read a single sheet. However it could be that
multiple sheets would hae been involved in getting
to the itemCount as provided.

The fix was to cleanup the doRead method and to use
the doRead method to read upto the correct index. To
prevent memory issues, the rowmapper is temporarily
being swapped for a noop mapper.

Fixes: #13