Handle exception when trying to read from file
Initially when loading a file there as a check if resource.getFile
would return null. However instead of returning null it will throw
a FileNotFoundException we need to handle.

With this commit we wrap the call in a try/catch and when an error
occurs we fallback to loading through an InputStream.

Fixes: #27